[Refactor] Split vllm_omni/entrypoints/cli/bench/serve.py Signed-off-by: ZacheryAU <zachery.au@gmail.com>
